api/requirements.txt
Django==1.9.4
Twisted==16.0.0
asgiref==0.10.0
autobahn==0.13.0
channels==0.10.2
daphne==0.10.0
decorator==4.0.9
django-braces==1.8.1
django-oauth-toolkit==0.10.0
djangorestframework==3.3.3
djangorestframework-jsonapi==2.0.0b2
inflection==0.3.1
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
txaio==2.2.2
zope.interface==4.1.3
